requirements.txt
absl-py==1.4.0
astor==0.8.1
cached-property==1.5.2
certifi==2021.5.30
cloudpickle==1.1.1
cycler==0.11.0
dataclasses==0.8
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
gast==0.2.2
google-pasta @ file:///Users/ktietz/demo/mc3/conda-bld/google-pasta_1630577991354/work
grpcio==1.48.2
h5py==3.1.0
importlib-metadata==4.8.3
joblib @ file:///tmp/build/80754af9/joblib_1613502643832/work
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy==1.19.5
opt-einsum @ file:///tmp/build/80754af9/opt_einsum_1621500238896/work
pandas==1.1.5
Pillow==8.4.0
protobuf==3.19.6
pyparsing==3.1.4
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1621365798935/work
scipy @ file:///tmp/build/80754af9/scipy_1597686625380/work
seaborn==0.11.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
tensorflow-probability==0.8.0rc0
termcolor==1.1.0
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
typing_extensions==4.1.1
webencodings==0.5.1
Werkzeug==2.0.3
wordcloud==1.9.2
wrapt==1.16.0
zipp==3.6.0
